Pipeline Alpha 2:
PreFreeSurfer:
•	Add gradient distortion correction module
•	Add MNI templates with the correct FOV for FSL
•	Add Mark’s new DO_mprage.sh script.
•	Make pipeline more generalizable to different kinds of data (e.g. data with only one average, data with no field map, data with a field map, data with reversed readout directions)
•	Add Mark’s new field map code and make it a global script instead of a module
•	Eliminate Code for –usesqform situations (which are eliminated with corrected template)
•	Gradient Distortion of Field Map

FreeSurfer:
•	Commit a change related to mris_make_surfaces
•	Don’t normalize T1w before pial surface
•	Re-Optimize pial surface placement based on T2 on a wider variety of data and newest mris_make_surfaces binary
•	Add a sqrt(T1w/T2w) ratio pial surface placement step

PostFreeSurfer:
•	Delete a bunch of unnecessary commands in CreateRibbon.sh
•	Change spec file distinguishers to .c5.spec and .wb.spec
•	Remove “external binary file” path for GIFTI files that are not external binary files
•	Improve native mesh/native volume spec files to include all data needed for evaluating myelin mapping/surface quality
•	Fix tags for native mesh/native volume surfaces
•	Improve all meshes spec files to include T2w image for other QA
•	Set volume palettes where needed
•	Import FreeSurfer label parcellations
•	Convert -metric-roi-mask, -metric-smoothing (dilate) to wb_command
•	Remove caret_command dependency for CreateRibbon.sh that is no longer needed
•	Add approximate FNIRT-based zblip correction for past MB scans

Generic Volume fMRI:
•	Add gradient distortion correction module
•	Gradient distortion correct the field map
•	Make a scout optional (it will be produced if NONE is specified).
•	Add Mark’s new field map code and make it a global script instead of a module
•	Add option for phase up phase down distortion correction (if I can get it to work now, or a placeholder if not)
•	Make sure it works with 1.6mm fMRI (or arbitrary resolution fMRI data)
•	Add gradient warp to one step resample
•	Fix issues with motion regressors not matching Wash U style (move translations in front, angles in degrees)
•	Incorporate AVI’s awk script for detrending motion regressors (as a separate ‘_dt.txt’ file)

Generic Surface fMRI:
•	Fix creation of zeros in brain ordinates
•	Eliminate Code for –usesqform situations
•	Use new wb_command functions from Tim Coalson to greatly speed up some processing steps.
•	Make sure it works with 1.6mm fMRI (or arbitrary resolution fMRI data)
•	Delete a bunch of unnecessary commands in RibbonVolumeToSurfaceMapping.sh
•	Move QA surface data onto 32k mesh for easier viewing

(Past two commits included in these changes)

+++ fMRIVolume/GenericfMRIVolumeProcessingPipeline.sh
@@ -15,13 +15,16 @@ InputNameOffMRI="$6"
 OutputNameOffMRI="$7"
 MagnitudeInputName="$8" #Expects 4D volume with two 3D timepoints
 PhaseInputName="$9"
-ScoutInputName="${10}"
+ScoutInputName="${10}" #Can be set to NONE, to fake, but this is not recommended
 DwellTime="${11}"
 TE="${12}"
 UnwarpDir="${13}"
-FinalFcMRIResolution="${14}"
+FinalfMRIResolution="${14}"
 PipelineScripts="${15}"
 GlobalScripts="${16}"
+DistortionCorrection="${17}" #FIELDMAP or TOPUP (not functional currently)
+GradientDistortionCoeffs="${18}"
+FNIRTConfig="${19}" #NONE to turn off approximate zblip correction
 
 #Naming Conventions
 T1wImage="T1w_acpc_dc"
@@ -33,69 +36,86 @@ ResultsFolder="Results"
 BiasField="BiasField_acpc_dc"
 BiasFieldMNI="BiasField"
 T1wAtlasName="T1w_restore"
-MovementRegressor="Movement_Regressors.txt"
+MovementRegressor="Movement_Regressors" #No extension, .txt appended
 MotionMatrixFolder="MotionMatrices"
 MotionMatrixPrefix="MAT_"
 FieldMapOutputName="FieldMap"
 MagnitudeOutputName="Magnitude"
 MagnitudeBrainOutputName="Magnitude_brain"
+ScoutName="Scout"
 FreeSurferBrainMask="brainmask_fs"
 fMRI2strOutputTransform="${OutputNameOffMRI}2str"
 RegOutput="Scout2T1w"
 AtlasTransform="acpc_dc2standard"
 OutputfMRI2StandardTransform="${OutputNameOffMRI}2standard"
+T2wRestoreImage="T2w_acpc_dc_restore"
+QAImage="T1wMulEPI"
 
-fMRIFolder="$fMRIFolder"
-FieldMapImageFolder="$FieldMapImageFolder"
-ScoutFolder="$ScoutFolder"
+fMRIFolder="$Path"/"$Subject"/"$fMRIFolder"
+FieldMapImageFolder="$Path"/"$Subject"/"$FieldMapImageFolder"
+ScoutFolder="$Path"/"$Subject"/"$ScoutFolder"
 T1wFolder="$Path"/"$Subject"/"$T1wFolder"
 AtlasSpaceFolder="$Path"/"$Subject"/"$AtlasSpaceFolder"
 ResultsFolder="$AtlasSpaceFolder"/"$ResultsFolder"/"$OutputNameOffMRI"
 
+#Create "Scout" if it doesn't exist
+if [ $ScoutInputName = "NONE" ] ; then
+  ScoutInputName="FakeScoutInput"
+  ScoutFolder="${fMRIFolder}_SB"
+  if [ ! -e $ScoutFolder ] ; then 
+    mkdir $ScoutFolder
+  fi
+  fslroi "$fMRIFolder"/"$InputNameOffMRI" "$fMRIFolder"_SB/"$ScoutInputName" 0 1
+fi
 
 #Gradient Distortion Correction of fMRI
-cp -r "$fMRIFolder"/"$InputNameOffMRI".nii.gz "$fMRIFolder"/"$OutputNameOffMRI".nii.gz
-
-#MotionCorrection
-##mkdir "$fMRIFolder"/MotionCorrection_mcFLIRTbased
-##"$PipelineComponents"/MotionCorrection_mcFLIRTbased.sh "$fMRIFolder"/MotionCorrection_mcFLIRTbased "$fMRIFolder"/"$OutputNameOffMRI" "$ScoutFolder"/"$ScoutInputName" "$fMRIFolder"/"$OutputNameOffMRI"_mc "$fMRIFolder"/"$MovementRegressor" "$fMRIFolder"/"$MotionMatrixFolder" "$MotionMatrixPrefix"
+if [ ! $GradientDistortionCoeffs = "NONE" ] ; then
+  if [ ! -e "$fMRIFolder"/GradientDistortionUnwarp ] ; then    
+    mkdir "$fMRIFolder"/GradientDistortionUnwarp
+  fi
+  cp "$fMRIFolder"/"$InputNameOffMRI" "$fMRIFolder"/GradientDistortionUnwarp/"$OutputNameOffMRI".nii.gz
+  "$GlobalScripts"/GradientDistortionUnwarp.sh "$fMRIFolder"/GradientDistortionUnwarp "$GradientDistortionCoeffs" "$fMRIFolder"/GradientDistortionUnwarp/"$OutputNameOffMRI" "$fMRIFolder"/"$OutputNameOffMRI"_gdc "$fMRIFolder"/"$OutputNameOffMRI"_gdc_warp
+  if [ ! -e "$ScoutFolder"/GradientDistortionUnwarp ] ; then    
+    mkdir "$ScoutFolder"/GradientDistortionUnwarp
+  fi
+  cp "$ScoutFolder"/"$ScoutInputName" "$ScoutFolder"/GradientDistortionUnwarp/"$ScoutName".nii.gz
+  "$GlobalScripts"/GradientDistortionUnwarp.sh "$ScoutFolder"/GradientDistortionUnwarp "$GradientDistortionCoeffs" "$ScoutFolder"/GradientDistortionUnwarp/"$ScoutName" "$ScoutFolder"/"$ScoutName"_gdc "$ScoutFolder"/"$ScoutName"_gdc_warp
+else
+  echo "NOT PERFORMING GRADIENT DISTORTION CORRECTION"
+  cp "$fMRIFolder"/"$InputNameOffMRI" "$fMRIFolder"/"$OutputNameOffMRI"_gdc.nii.gz
+  fslroi "$fMRIFolder"/"$OutputNameOffMRI"_gdc.nii.gz "$fMRIFolder"/"$OutputNameOffMRI"_gdc_warp.nii.gz 0 3
+  fslmaths "$fMRIFolder"/"$OutputNameOffMRI"_gdc_warp.nii.gz -mul 0 "$fMRIFolder"/"$OutputNameOffMRI"_gdc_warp.nii.gz
+  cp "$ScoutFolder"/"$ScoutInputName" "$ScoutFolder"/"$ScoutName"_gdc.nii.gz
+fi
 
 mkdir -p "$fMRIFolder"/MotionCorrection_FLIRTbased
-"$PipelineScripts"/MotionCorrection_FLIRTbased.sh "$fMRIFolder"/MotionCorrection_FLIRTbased "$fMRIFolder"/"$OutputNameOffMRI" "$ScoutFolder"/"$ScoutInputName" "$fMRIFolder"/"$OutputNameOffMRI"_mc "$fMRIFolder"/"$MovementRegressor" "$fMRIFolder"/"$MotionMatrixFolder" "$MotionMatrixPrefix" "$PipelineScripts" "$GlobalScripts"
-
-#FieldMap Preprocessing
-mkdir -p "$fMRIFolder"/FieldMapPreProcessing
-"$GlobalScripts"/FieldMapPreProcessing.sh "$fMRIFolder"/FieldMapPreProcessing "$FieldMapImageFolder"/"$MagnitudeInputName" "$FieldMapImageFolder"/"$PhaseInputName" "$FieldMapOutputName" "$MagnitudeOutputName" "$MagnitudeBrainOutputName" "$TE"
-
-#EPI Distortion Correction and EPI to T1w Registration (For preprocessing evaluation, not for main pipeline)
-##mkdir "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRbased
-##"$PipelineComponents"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRbased.sh "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRbased "$fMRIFolder"/"$OutputNameOffMRI"_mc "$ScoutFolder"/"$ScoutInputName" "$T1wFolder"/"$T1wImage" "$T1wFolder"/"$T1wRestoreImage" "$T1wFolder"/"$FreeSurferBrainMask" "$FieldMapImageFolder"/"$FieldMapOutputName" "$FieldMapImageFolder"/"$MagnitudeOutputName" "$FieldMapImageFolder"/"$MagnitudeBrainOutputName" "$DwellTime" "$UnwarpDir" "$T1wFolder"/xfms/"$fMRI2strOutputTransform" "$T1wFolder"/"$BiasField" "$fMRIFolder"/"$RegOutput"
-
-#EPI Distortion Correction and EPI to T1w Registration (For preprocessing evaluation, not for main pipeline)
-##mkdir "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FugueAndFreeSurferBBRbased
-##"$PipelineComponents"/DistortionCorrectionAndEPIToT1wReg_FugueAndFreeSurferBBRbased.sh "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FugueAndFreeSurferBBRbased "$fMRIFolder"/"$OutputNameOffMRI"_mc "$ScoutFolder"/"$ScoutInputName" "$T1wFolder"/"$T1wImage" "$T1wFolder"/"$T1wRestoreImage" "$T1wFolder"/"$FreeSurferBrainMask" "$FieldMapImageFolder"/"$FieldMapOutputName" "$FieldMapImageFolder"/"$MagnitudeOutputName" "$FieldMapImageFolder"/"$MagnitudeBrainOutputName" "$DwellTime" "$UnwarpDir" "$T1wFolder"/xfms/"$fMRI2strOutputTransform" "$T1wFolder"/"$BiasField" "$fMRIFolder"/"$RegOutput" "$T1wFolder" "$Subject" 
-
-#EPI Distortion Correction and EPI to T1w Registration (For preprocessing evaluation, not for main pipeline) includes VSM option (Unfinished)
-##mkdir "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FugueAndFreeSurferBBRbasedVSM
-##"$PipelineComponents"/DistortionCorrectionAndEPIToT1wReg_FugueAndFreeSurferBBRbasedVSM.sh "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FugueAndFreeSurferBBRbasedVSM "$fMRIFolder"/"$OutputNameOffMRI"_mc "$ScoutFolder"/"$ScoutInputName" "$T1wFolder"/"$T1wImage" "$T1wFolder"/"$T1wRestoreImage" "$T1wFolder"/"$FreeSurferBrainMask" "$FieldMapImageFolder"/"$FieldMapOutputName" "$FieldMapImageFolder"/"$MagnitudeOutputName" "$FieldMapImageFolder"/"$MagnitudeBrainOutputName" "$DwellTime" "$UnwarpDir" "$T1wFolder"/xfms/"$fMRI2strOutputTransform" "$T1wFolder"/"$BiasField" "$fMRIFolder"/"$RegOutput" "$T1wFolder" "$Subject" 
+"$PipelineScripts"/MotionCorrection_FLIRTbased.sh "$fMRIFolder"/MotionCorrection_FLIRTbased "$fMRIFolder"/"$OutputNameOffMRI"_gdc "$ScoutFolder"/"$ScoutName"_gdc "$fMRIFolder"/"$OutputNameOffMRI"_mc "$fMRIFolder"/"$MovementRegressor" "$fMRIFolder"/"$MotionMatrixFolder" "$MotionMatrixPrefix" "$PipelineScripts" "$GlobalScripts"
 
 #EPI Distortion Correction and EPI to T1w Registration
-if [ -e "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased ] ; then
-  rm -r "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased
+if [ $DistortionCorrection = "FIELDMAP" ] ; then
+  if [ -e "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased ] ; then
+    rm -r "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased
+  fi
+  mkdir -p "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased
+  "$PipelineScripts"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased.sh "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased "$ScoutFolder"/"$ScoutName"_gdc "$T1wFolder"/"$T1wImage" "$T1wFolder"/"$T1wRestoreImage" "$T1wFolder"/"$T1wRestoreImageBrain" "$FieldMapImageFolder"/"$MagnitudeInputName" "$FieldMapImageFolder"/"$PhaseInputName" "$TE" "$DwellTime" "$UnwarpDir" "$T1wFolder"/xfms/"$fMRI2strOutputTransform" "$T1wFolder"/"$BiasField" "$fMRIFolder"/"$RegOutput" "$T1wFolder" "$Subject" "$GlobalScripts" "$GradientDistortionCoeffs" "$T1wFolder"/"$T2wRestoreImage" "$FNIRTConfig" "$fMRIFolder"/"$QAImage"
+elif [ $DistortionCorrection = "TOPUP" ] ; then
+  echo "TOPUP fMRI DISTORTION CORRECTION NOT YET IMPLEMENTED"
+  exit
+else
+  echo "UNKNOWN DISTORTION CORRECTION METHOD"
+  exit
 fi
-mkdir -p "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased
-
-"$PipelineScripts"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased.sh "$fMRIFolder"/DistortionCorrectionAndEPIToT1wReg_FLIRTBBRAndFreeSurferBBRbased "$ScoutFolder"/"$ScoutInputName" "$T1wFolder"/"$T1wImage" "$T1wFolder"/"$T1wRestoreImage" "$T1wFolder"/"$T1wRestoreImageBrain" "$fMRIFolder"/FieldMapPreProcessing/"$FieldMapOutputName" "$fMRIFolder"/FieldMapPreProcessing/"$MagnitudeOutputName" "$fMRIFolder"/FieldMapPreProcessing/"$MagnitudeBrainOutputName" "$DwellTime" "$UnwarpDir" "$T1wFolder"/xfms/"$fMRI2strOutputTransform" "$T1wFolder"/"$BiasField" "$fMRIFolder"/"$RegOutput" "$T1wFolder" "$Subject" "$GlobalScripts"
 
 #One Step Resampling
 mkdir -p "$fMRIFolder"/OneStepResampling
-"$PipelineScripts"/OneStepResampling.sh "$fMRIFolder"/OneStepResampling "$fMRIFolder"/"$OutputNameOffMRI" "$AtlasSpaceFolder"/"$T1wAtlasName" "$FinalFcMRIResolution" "$AtlasSpaceFolder" "$T1wFolder"/xfms/"$fMRI2strOutputTransform" "$AtlasSpaceFolder"/xfms/"$AtlasTransform" "$AtlasSpaceFolder"/xfms/"$OutputfMRI2StandardTransform" "$fMRIFolder"/"$MotionMatrixFolder" "$MotionMatrixPrefix" "$fMRIFolder"/"$OutputNameOffMRI"_nonlin "$AtlasSpaceFolder"/"$FreeSurferBrainMask" "$AtlasSpaceFolder"/"$BiasFieldMNI"
+"$PipelineScripts"/OneStepResampling.sh "$fMRIFolder"/OneStepResampling "$fMRIFolder"/"$InputNameOffMRI" "$AtlasSpaceFolder"/"$T1wAtlasName" "$FinalfMRIResolution" "$AtlasSpaceFolder" "$T1wFolder"/xfms/"$fMRI2strOutputTransform" "$AtlasSpaceFolder"/xfms/"$AtlasTransform" "$AtlasSpaceFolder"/xfms/"$OutputfMRI2StandardTransform" "$fMRIFolder"/"$MotionMatrixFolder" "$MotionMatrixPrefix" "$fMRIFolder"/"$OutputNameOffMRI"_nonlin "$AtlasSpaceFolder"/"$FreeSurferBrainMask" "$AtlasSpaceFolder"/"$BiasFieldMNI" "$fMRIFolder"/"$OutputNameOffMRI"_gdc_warp
 
 #Intensity Normalization and Bias Removal
-"$PipelineScripts"/IntensityNormalization.sh "$fMRIFolder"/"$OutputNameOffMRI"_nonlin "$AtlasSpaceFolder"/"$BiasFieldMNI"."$FinalFcMRIResolution" "$AtlasSpaceFolder"/"$FreeSurferBrainMask"."$FinalFcMRIResolution" "$fMRIFolder"/"$OutputNameOffMRI"_nonlin_norm
+"$PipelineScripts"/IntensityNormalization.sh "$fMRIFolder"/"$OutputNameOffMRI"_nonlin "$AtlasSpaceFolder"/"$BiasFieldMNI"."$FinalfMRIResolution" "$AtlasSpaceFolder"/"$FreeSurferBrainMask"."$FinalfMRIResolution" "$fMRIFolder"/"$OutputNameOffMRI"_nonlin_norm
 
 mkdir -p "$ResultsFolder"
 cp -r "$fMRIFolder"/"$OutputNameOffMRI"_nonlin_norm.nii.gz "$ResultsFolder"/"$OutputNameOffMRI".nii.gz
-cp -r "$fMRIFolder"/"$MovementRegressor" "$ResultsFolder"/"$MovementRegressor"
+cp -r "$fMRIFolder"/"$MovementRegressor".txt "$ResultsFolder"/"$MovementRegressor".txt
+cp -r "$fMRIFolder"/"$MovementRegressor"_dt.txt "$ResultsFolder"/"$MovementRegressor"_dt.txt
